Geodesy types are no longer subtypes of FixedVectorNoTuple
